# docs/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QUEST API",
    "version": "0.0.0-dev",
    "description": "Generated from the zod contracts in @quest/types (ADR-012). Conventions: docs/api/API_CONVENTIONS.md. Regenerate with `pnpm --filter @quest/api openapi:generate`."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "local"
    }
  ],
...
  "paths": {
...
    "/v1/auth/email/resend": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Resend the verification code",
        "operationId": "post-auth-email-resend",
        "parameters": [],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accepted": {
                      "type": "boolean",
                      "const": true
                    }
                  },
                  "required": [
                    "accepted"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION"
        ],
        "x-quest-rate-limit": {
          "limit": 3,
          "windowSeconds": 300
        }
      }
    },
    "/v1/auth/email/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify the email address with the mailed code",
        "operationId": "post-auth-email-verify",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accountId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "email": {
                      "type": "string"
                    },
                    "emailVerified": {
                      "type": "boolean"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "PENDING_VERIFICATION",
                        "ACTIVE",
                        "DEACTIVATED",
                        "SUSPENDED",
                        "DELETION_REQUESTED",
                        "DELETED"
                      ]
                    },
                    "ageBand": {
                      "type": "string",
                      "enum": [
                        "TEEN_13_15",
                        "TEEN_16_17",
                        "ADULT"
                      ]
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "USER",
                          "SUPER_ADMIN",
                          "TRUST_SAFETY_LEAD",
                          "MODERATOR",
                          "SUPPORT",
                          "ANALYST",
                          "READ_ONLY"
                        ]
                      }
                    },
                    "onboarding": {
                      "type": "object",
                      "properties": {
                        "completed": {
                          "type": "boolean"
                        },
                        "nextStep": {
                          "type": "string",
                          "enum": [
                            "VERIFY_EMAIL",
                            "PROFILE",
                            "INTERESTS",
                            "DONE"
                          ]
                        }
                      },
                      "required": [
                        "completed",
                        "nextStep"
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    "deletionScheduledFor": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "accountId",
                    "email",
                    "emailVerified",
                    "state",
                    "ageBand",
                    "roles",
                    "onboarding",
                    "createdAt",
                    "deletionScheduledFor"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION"
        ],
        "x-quest-rate-limit": {
          "limit": 10,
          "windowSeconds": 60
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "pattern": "^[0-9]{6}$"
                  }
                },
                "required": [
                  "code"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in with email and password",
        "operationId": "post-auth-login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "account": {
                      "type": "object",
                      "properties": {
                        "accountId": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "email": {
                          "type": "string"
                        },
                        "emailVerified": {
                          "type": "boolean"
                        },
                        "state": {
                          "type": "string",
                          "enum": [
                            "PENDING_VERIFICATION",
                            "ACTIVE",
                            "DEACTIVATED",
                            "SUSPENDED",
                            "DELETION_REQUESTED",
                            "DELETED"
                          ]
                        },
                        "ageBand": {
                          "type": "string",
                          "enum": [
                            "TEEN_13_15",
                            "TEEN_16_17",
                            "ADULT"
                          ]
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "USER",
                              "SUPER_ADMIN",
                              "TRUST_SAFETY_LEAD",
                              "MODERATOR",
                              "SUPPORT",
                              "ANALYST",
                              "READ_ONLY"
                            ]
                          }
                        },
                        "onboarding": {
                          "type": "object",
                          "properties": {
                            "completed": {
                              "type": "boolean"
                            },
                            "nextStep": {
                              "type": "string",
                              "enum": [
                                "VERIFY_EMAIL",
                                "PROFILE",
                                "INTERESTS",
                                "DONE"
                              ]
                            }
                          },
                          "required": [
                            "completed",
                            "nextStep"
                          ]
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        "deletionScheduledFor": {
                          "anyOf": [
                            {
                              "type": "string",
                              "format": "date-time",
                              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      },
                      "required": [
                        "accountId",
                        "email",
                        "emailVerified",
                        "state",
                        "ageBand",
                        "roles",
                        "onboarding",
                        "createdAt",
                        "deletionScheduledFor"
                      ]
                    },
                    "tokens": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "minLength": 1
                        },
                        "accessTokenExpiresIn": {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991
                        },
                        "refreshToken": {
                          "type": "string",
                          "minLength": 1
                        },
                        "refreshTokenExpiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        "sessionId": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "tokenType": {
                          "type": "string",
                          "const": "Bearer"
                        }
                      },
                      "required": [
                        "accessToken",
                        "accessTokenExpiresIn",
                        "refreshToken",
                        "refreshTokenExpiresAt",
                        "sessionId",
                        "tokenType"
                      ]
                    }
                  },
                  "required": [
                    "account",
                    "tokens"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [],
        "x-quest-permissions": [],
        "x-quest-account-states": [],
        "x-quest-rate-limit": {
          "limit": 10,
          "windowSeconds": 60
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "client": {
                    "type": "object",
                    "properties": {
                      "platform": {
                        "type": "string",
                        "enum": [
                          "IOS",
                          "ANDROID",
                          "WEB"
                        ]
                      },
                      "installationId": {
                        "type": "string",
                        "pattern": "^[A-Za-z0-9._-]{8,128}$"
                      },
                      "appVersion": {
                        "type": "string",
                        "maxLength": 32
                      },
                      "deviceName": {
                        "type": "string",
                        "maxLength": 64
                      }
                    },
                    "required": [
                      "platform",
                      "installationId"
                    ]
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign out the current session",
        "operationId": "post-auth-logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION",
          "DELETION_REQUESTED"
        ]
      }
    },
    "/v1/auth/logout-all": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign out everywhere",
        "operationId": "post-auth-logout-all",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revokedSessions": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "revokedSessions"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION",
          "DELETION_REQUESTED"
        ]
      }
    },
    "/v1/auth/password/change": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change password (revokes other sessions)",
        "operationId": "post-auth-password-change",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission, disallowed account state, or unverified email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-quest-permissions": [],
        "x-quest-account-states": [
          "ACTIVE",
          "PENDING_VERIFICATION"
        ],
        "x-quest-rate-limit": {
          "limit": 5,
          "windowSeconds": 60
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 128
                  }
                },
                "required": [
                  "currentPassword",
                  "newPassword"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/auth/password/forgot": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request a password-reset code (always 202)",
        "operationId": "post-auth-password-forgot",
        "parameters": [],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accepted": {
                      "type": "boolean",
                      "const": true
                    }
                  },
                  "required": [
                    "accepted"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [],
        "x-quest-permissions": [],
        "x-quest-account-states": [],
        "x-quest-rate-limit": {
          "limit": 3,
          "windowSeconds": 300
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/auth/password/reset": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Reset the password with the mailed code",
        "operationId": "post-auth-password-reset",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (retry-after header)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [],
        "x-quest-permissions": [],
        "x-quest-account-states": [],
        "x-quest-rate-limit": {
          "limit": 5,
          "windowSeconds": 60
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^[0-9]{6}$"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 128
                  }
                },
                "required": [
                  "email",
                  "code",
                  "newPassword"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "PAYLOAD_TOO_LARGE",
                  "UNSUPPORTED_MEDIA_TYPE",
                  "SERVICE_UNAVAILABLE",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "issues": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ]
                }
              },
              "correlationId": {
                "type": "string"
              },
              "timestamp": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "correlationId",
              "timestamp"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}